refactor: split SszEx into different submodules
