+++ bench/ssz_ex_bench.exs
@@ -1,4 +1,6 @@
-alias LambdaEthereumConsensus.SszEx
+alias LambdaEthereumConsensus.SszEx.Decode
+alias LambdaEthereumConsensus.SszEx.Encode
+alias LambdaEthereumConsensus.SszEx.Merkleization
 alias LambdaEthereumConsensus.Store.StateDb
 alias Types.BeaconState
 
@@ -11,7 +13,7 @@ alias Types.BeaconState
 
 Benchee.run(
   %{
-    "SszEx.decode" => fn -> SszEx.decode(encoded_state, BeaconState) end,
+    "Decode.decode" => fn -> Decode.decode(encoded_state, BeaconState) end,
     "Ssz.from_ssz" => fn -> Ssz.from_ssz(encoded_state, BeaconState) end
   },
   warmup: 2,
@@ -20,7 +22,7 @@ Benchee.run(
 
 Benchee.run(
   %{
-    "SszEx.encode" => fn -> SszEx.encode(state, BeaconState) end,
+    "Encode.encode" => fn -> Encode.encode(state, BeaconState) end,
     "Ssz.to_ssz" => fn -> Ssz.to_ssz(state) end
   },
   warmup: 2,
@@ -29,7 +31,9 @@ Benchee.run(
 
 Benchee.run(
   %{
-    "SszEx.hash_tree_root!" => fn -> SszEx.hash_tree_root!(state, BeaconState) end,
+    "Merkleization.hash_tree_root!" => fn ->
+      Merkleization.hash_tree_root!(state, BeaconState)
+    end,
     "Ssz.hash_tree_root" => fn -> Ssz.hash_tree_root(state, BeaconState) end
   },
   warmup: 2,
