+++ lib/lambda_ethereum_consensus/validator/utils.ex
@@ -2,7 +2,7 @@ defmodule LambdaEthereumConsensus.Validator.Utils do
   @moduledoc """
   Functions for performing validator duties.
   """
-  alias LambdaEthereumConsensus.SszEx
+  alias LambdaEthereumConsensus.SszEx.Hash
   alias LambdaEthereumConsensus.StateTransition.Accessors
   alias LambdaEthereumConsensus.StateTransition.Misc
   alias Types.AttestationData
@@ -107,7 +107,7 @@ defmodule LambdaEthereumConsensus.Validator.Utils do
     target = Constants.target_aggregators_per_committee()
     modulo = committee_length |> div(target) |> max(1)
 
-    SszEx.hash(slot_signature)
+    Hash.hash(slot_signature)
     |> binary_part(0, 8)
     |> :binary.decode_unsigned(:little)
     |> rem(modulo) == 0
